requirements.txt
PyYAML==5.3.1
py-zabbix==1.1.7
